LPD-93353 Move FormulaContainer and TranslationOptionsContainer specs under Tabs/BasicInfo
