+++ modules/apps/object/object-web/src/main/resources/META-INF/resources/js/tests/ObjectField/Tabs/BasicInfo/FormulaContainer.spec.tsx
@@ -7,7 +7,7 @@ import '@testing-library/jest-dom';
 import {render, screen} from '@testing-library/react';
 import React from 'react';
 
-import {FormulaContainer} from '../../components/ObjectField/Tabs/BasicInfo/FormulaContainer';
+import {FormulaContainer} from '../../../../components/ObjectField/Tabs/BasicInfo/FormulaContainer';
 
 const defaultProps = {
 	errors: {},
